+++ src/core/baseClass/Epic.js
@@ -1,0 +1,5 @@
+export default class Epic {
+	// 类型是epic
+	$type = 'epic';
+	
+}