+++ src/page/room/app/index.js
@@ -2,13 +2,15 @@
  *  启动器创建模块 app 的入口文件。传入的组件必须是模块的根组件。
  */
 
+import { lazy } from 'react'
 import Launcher from '@home/core/launcher';
 import loggerMiddleware from '@home/middlewares/logger';
 
 import epics from '../epics';
 import reducers from '../controllers';
+// import PageEntry from '../modules';
 
-import PageEntry from '../modules';
+const PageEntry = lazy(() => import('../modules'))
 
 export default Launcher.createApp({
     name: 'roomApp',    //  应用名称，每个应用唯一