+++ src/page/room/epics/demo.epic.js
@@ -2,10 +2,11 @@ import { ofType } from 'redux-observable'
 import { filter, mapTo, mergeMap, map } from 'rxjs/operators'
 import http$ from '@home/util/http'
 import Event from '@home/core/event'
+import Epic from '@home/core/baseClass/Epic';
 
 const event = new Event()
 
-export default class ListEpic {
+export default class ListEpic extends Epic{
 	// epic命名空间,与controller ns 对应
 	namespace = 'list';